Move Portal component outside of Draftail
