+++ examples/blocks/MediaBlock.js
@@ -2,9 +2,10 @@ import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { EditorState, Modifier, SelectionState } from 'draft-js';
 
-import { Icon, Portal } from '../../lib';
+import { Icon } from '../../lib';
 
 import Tooltip from '../components/Tooltip';
+import Portal from '../components/Portal';
 
 // Defined in both CSS and JS to constraint the maximum size of the tooltip.
 const OPTIONS_MAX_WIDTH = 300;
