+++ examples/entities/Link.js
@@ -1,10 +1,12 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 
-import { Icon, Portal } from '../../lib';
+import { Icon } from '../../lib';
+
+import { truncateURL } from '../utils/format';
 
 import Tooltip from '../components/Tooltip';
-import { truncateURL } from '../utils/format';
+import Portal from '../components/Portal';
 
 class Link extends Component {
     constructor(props) {
