+++ lib/index.js
@@ -1,6 +1,5 @@
 import DraftailEditor from './components/DraftailEditor';
 import Icon from './components/Icon';
-import Portal from './components/Portal';
 import DraftUtils from './api/DraftUtils';
 import { BLOCK_TYPE, ENTITY_TYPE, INLINE_STYLE } from './api/constants';
 
@@ -11,4 +10,4 @@ import { BLOCK_TYPE, ENTITY_TYPE, INLINE_STYLE } from './api/constants';
 
 export default DraftailEditor;
 
-export { DraftUtils, BLOCK_TYPE, ENTITY_TYPE, INLINE_STYLE, Icon, Portal };
+export { DraftUtils, BLOCK_TYPE, ENTITY_TYPE, INLINE_STYLE, Icon };
